feature: adicionando remover anuncios no app
