+++ app/build.gradle
@@ -17,7 +17,7 @@ android {
         applicationId "com.rodrigmatrix.sigaaufc"
         minSdkVersion 21
         targetSdkVersion 29
-        versionCode 21
+        versionCode 22
         versionName "3.1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -53,7 +53,7 @@ dependencies {
     def lifecycle_version = '2.2.0-alpha02'
     def app_compact_version = '1.1.0'
     def constraint_layout_version = '1.1.3'
-    def core_version = '1.1.0'
+    def core_version = '1.2.0'
     def anko_version = '0.10.8'
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
@@ -64,13 +64,13 @@ dependencies {
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
 
     // Firebase
-    implementation 'com.google.firebase:firebase-core:17.2.1'
+    implementation 'com.google.firebase:firebase-core:17.2.2'
     implementation 'com.google.firebase:firebase-ads:18.3.0'
-    implementation 'com.google.firebase:firebase-inappmessaging-display:19.0.2'
+    implementation 'com.google.firebase:firebase-inappmessaging-display:19.0.3'
     implementation 'com.google.firebase:firebase-messaging:20.1.0'
-    implementation 'com.google.firebase:firebase-analytics:17.2.1'
-    implementation 'com.google.firebase:firebase-config:19.1.0'
-    implementation 'com.google.android.play:core:1.6.4'
+    implementation 'com.google.firebase:firebase-analytics:17.2.2'
+    implementation 'com.google.firebase:firebase-config:19.1.1'
+    implementation 'com.google.android.play:core:1.6.5'
     implementation('com.crashlytics.sdk.android:crashlytics:2.10.1') {
         transitive = true
     }
@@ -90,9 +90,9 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.1.0'
 
+
     implementation 'com.github.igorronner:IRAds:1.3.7'
-
-
+    implementation 'com.android.billingclient:billing:1.1'
 
     kapt "androidx.room:room-compiler:$room_version"
 
@@ -101,20 +101,22 @@ dependencies {
 
     // Kotlin Coroutines
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.3'
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.2"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.3"
 
     // Glide
     implementation 'com.github.bumptech.glide:glide:4.10.0'
     kapt 'com.github.bumptech.glide:compiler:4.10.0'
 
     // Okhttp
+    implementation 'com.squareup.retrofit2:retrofit:2.6.2'
     implementation "com.squareup.okhttp3:okhttp:4.0.1"
+    implementation 'com.github.franmontiel:PersistentCookieJar:v1.0.1'
 
     // Preference
     implementation "androidx.preference:preference:1.1.0"
 
     // Material Components
-    implementation "com.google.android.material:material:1.2.0-alpha03"
+    implementation "com.google.android.material:material:1.2.0-alpha05"
 
 
     // Jsoup
