+++ app/src/main/res/layout/file_row.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:gravity="center"
@@ -24,7 +25,7 @@
                 android:orientation="vertical">
 
             <TextView
-                    android:text="Nome do arquivo"
+                    tools:text="Nome do arquivo"
                     android:gravity="center"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -40,6 +41,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:text="Download"
+                    app:cornerRadius="8dp"
                     android:layout_marginEnd="8dp"
                     android:layout_gravity="end|bottom"
                     android:textColorLink="@color/colorPrimary"
