+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -20,6 +20,7 @@
             android:id="@+id/nav_library"
             android:icon="@drawable/ic_local_library"
             android:title="@string/library_shortcut" />
+
         <item
                 android:id="@+id/nav_settings"
                 android:icon="@drawable/ic_settings"
